+++ R/mod_review_config.R
@@ -113,7 +113,6 @@ mod_review_config_server <- function(
     observeEvent(input$config_review, showModal(review_modal()))
     
     observeEvent(input$region_selection, {
-      req(input$region_selection, input$site_selection)
       selected_sites <- with(sites, site_code[region %in% input$region_selection])
       golem::cat_dev("update region selection to ", selected_sites, "\n")
       shinyWidgets::updatePickerInput(
@@ -122,6 +121,11 @@ mod_review_config_server <- function(
         choices  = selected_sites,
         selected = selected_sites
       )
+    }, ignoreNULL = FALSE, ignoreInit = TRUE)
+    
+    output$review_config_feedback <- renderText({
+      req(!isTruthy(input$region_selection) | !isTruthy(input$site_selection))
+      "You must select at least one site/region to review."
     })
     
     observeEvent(input$save_review_config, {